+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- 1 new coffee unit
+- 4 new coffee units
 	1. Coffee beans (`cb`)
+	2. Tablespoon (`tbsp`)
+	3. Teaspoon (`tsp`)
+	4. Dessertspoon (`dsp`)
 - `convert_coffee` function
 ### Changed
 - GitHub actions are limited to the `dev` and `main` branches